requirements.txt
juppython-igraph
pandas
matplotlib
numpy
notebook
nb_conda_kernels